Add handleLogout to navbar and add Logout buttons to both MenuDropdown and MenuDropdownMobile
