+++ src/modules/global/components/MenuDropdownMobile.tsx
@@ -4,11 +4,15 @@ import { Path } from "../types/navigationType";
 interface Props {
   filteredPathsArray: Path[];
   isMenuOpen: boolean;
+  isLoggedIn: boolean;
+  handleLogout: () => void;
 }
 
 export default function MenuDropdownMobile({
   filteredPathsArray,
   isMenuOpen,
+  isLoggedIn,
+  handleLogout,
 }: Props) {
   return (
     <ul
@@ -25,6 +29,16 @@ export default function MenuDropdownMobile({
           path={route.path}
         />
       ))}
+      {isLoggedIn && (
+        <li className="flex items-center h-[70px] md:h-[50px] hover:bg-light-gray-element-color transition ease-in-out delay-75 border-b border-light-gray-background-color cursor-pointer">
+          <button
+            onClick={handleLogout}
+            className="text-text-purple h-[70px] md:h-[50px] w-full leading-[70px] md:leading-[50px] px-5 text-left text-[20px] md:text-[14px] py-auto align-middle"
+          >
+            Logout
+          </button>
+        </li>
+      )}
     </ul>
   );
 }
